Phase 17: Code Review Hardening - transaction atomicity, double serialization fix, SecretStr, prompt injection protection, fallback safety, schema type tightening, CORS tightening, frontend timeout protection
- Fix transaction atomicity: workspace/invitation/agent_flow/assignment services use flush+single commit, sub-services accept auto_commit=False
- Fix double JSON serialization in member_profile/risk/project services
- SecretStr for llm_api_key, PositiveFloat for llm_timeout_seconds
- XML tag isolation for workspace_state in prompts (injection protection)
- Fallback payloads return empty lists instead of None IDs
- Schema type tightening: skills/evidence/dependency_ids/acceptance_criteria/done_criteria
- Global exception handler, CORS tightened, lifespan error handling
- Database session explicit rollback on exception
- JSON repair layered approach in workflow
- LLMError propagation to fallback path
- Frontend request<T> AbortController timeout + JSON parse protection
- Extract require_row() to core/db_utils.py
- Invitation accept with user_id parameter
- Update AGENTS.md/CLAUDE.md/handoff.md with new patterns and Phase 17

+++ backend/app/models/member_profile.py
@@ -11,7 +11,7 @@ class MemberProfile(SQLModel, table=True):
     user_id: str = Field(foreign_key="users.id")
     workspace_id: str = Field(foreign_key="workspaces.id")
     skills: str = Field(default="[]")  # JSON string: [{"name": "frontend", "level": 3}]
-    available_hours_per_week: int = Field(default=0)
+    available_hours_per_week: float = Field(default=0.0)
     role_preference: str = Field(default="")
     interests: str = Field(default="")
     constraints: str = Field(default="")
